完善缺點

+++ backend-Spingboot/src/main/java/com/example/movie/model/User.java
@@ -5,19 +5,43 @@ import jakarta.persistence.*;
 @Entity
 @Table(name = "users")
 public class User {
-    @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Integer id;
     private String username;
     private String password;
     private String role;
 
     // getters/setters
-    public Long getId(){return id;}
-    public void setId(Long id){this.id = id;}
-    public String getUsername(){return username;}
-    public void setUsername(String u){this.username=u;}
-    public String getPassword(){return password;}
-    public void setPassword(String p){this.password=p;}
-    public String getRole(){return role;}
-    public void setRole(String r){this.role=r;}
+    public Integer getId() {
+        return id;
+    }
+
+    public void setId(Integer id) {
+        this.id = id;
+    }
+
+    public String getUsername() {
+        return username;
+    }
+
+    public void setUsername(String u) {
+        this.username = u;
+    }
+
+    public String getPassword() {
+        return password;
+    }
+
+    public void setPassword(String p) {
+        this.password = p;
+    }
+
+    public String getRole() {
+        return role;
+    }
+
+    public void setRole(String r) {
+        this.role = r;
+    }
 }
